+++ sites/partners/src/listings/PaperListingDetails/sections/DetailRankingsAndResults.tsx
@@ -1,10 +1,10 @@
 import React, { useContext } from "react"
 import moment from "moment"
-import { t, GridSection, ViewItem } from "@bloom-housing/ui-components"
+import { t, GridSection, ViewItem, GridCell } from "@bloom-housing/ui-components"
 import { ListingContext } from "../../ListingContext"
 import { getLotteryEvent } from "../../../../lib/helpers"
 import { EnumListingReviewOrderType } from "@bloom-housing/backend-core/types"
-import { getDetailFieldNumber } from "./helpers"
+import { getDetailFieldNumber, getDetailFieldString } from "./helpers"
 
 const DetailRankingsAndResults = () => {
   const listing = useContext(ListingContext)
@@ -81,6 +81,13 @@ const DetailRankingsAndResults = () => {
           </ViewItem>
         </GridSection>
       )}
+      <GridSection columns={1}>
+        <GridCell>
+          <ViewItem label={t("listings.whatToExpectLabel")}>
+            {getDetailFieldString(listing.whatToExpect)}
+          </ViewItem>
+        </GridCell>
+      </GridSection>
     </GridSection>
   )
 }