+++ CHANGELOG.md
@@ -47,6 +47,7 @@ All notable changes to this project will be documented in this file. The format 
   - removed roles for public users and assigned a "partner" role for leasing agents([#1628](https://github.com/bloom-housing/bloom/pull/1628))
   - Updates redis reset call to flush all keys
   - Updated listing's importer to handle latest reserved community type changes ([#1667](https://github.com/bloom-housing/bloom/pull/1667)) (Emily Jablonski)
+  - Change whatToExpect to be a string instead of a json blob, make it editable in listings management ([#1681](https://github.com/bloom-housing/bloom/pull/1681)) (Emily Jablonski)
 
 - Fixed:
   - Added checks for property in listing.dto transforms